+++ api/rest-client.ts
@@ -153,8 +153,11 @@ export class RingRestClient {
             ? 'refresh token is'
             : 'email and password are',
         errorMessage =
-          `Failed to fetch oauth token from Ring. Verify that your ${authTypeMessage} correct. ` +
-          responseError
+          'Failed to fetch oauth token from Ring. ' +
+          (responseData.err_msg === 'too many requests from dependency service'
+            ? 'You have requested too many 2fa codes.  Ring limits 2fa to 10 codes within 10 minutes.  Please try again in 10 minutes.'
+            : `Verify that your ${authTypeMessage} correct.`) +
+          ` (error: ${responseError})`
       logError(requestError.response || requestError)
       logError(errorMessage)
       throw new Error(errorMessage)
@@ -227,7 +230,8 @@ export class RingRestClient {
     json?: boolean
     responseType?: ResponseType
   }): Promise<T & ExtendedResponse> {
-    const { method, url, data, json, responseType } = options
+    const { method, url, data, json, responseType } = options,
+      hardwareId = await hardwareIdPromise
 
     try {
       await this.sessionPromise
@@ -237,7 +241,7 @@ export class RingRestClient {
             ? 'application/json'
             : 'application/x-www-form-urlencoded',
           authorization: `Bearer ${authTokenResponse.access_token}`,
-          hardware_id: await hardwareIdPromise
+          hardware_id: hardwareId
         }
 
       return await requestWithRetry<T>({
@@ -275,13 +279,15 @@ export class RingRestClient {
           return this.request(options)
         }
         logError(
-          `http request failed.  ${url} returned unknown errors: (${errors}).`
+          `http request failed.  ${url} returned unknown errors: (${stringify(
+            errors
+          )}).`
         )
       }
 
       if (response.status === 404 && url.startsWith(clientApiBaseUrl)) {
         logError('404 from endpoint ' + url)
-        if (response.data === '') {
+        if (response.data?.error?.includes(hardwareId)) {
           logError(
             'Session hardware_id not found.  Creating a new session and trying again.'
           )
@@ -292,7 +298,11 @@ export class RingRestClient {
         throw new Error('Not found with response: ' + stringify(response.data))
       }
 
-      logError(`Request to ${url} failed`)
+      logError(
+        `Request to ${url} failed with status ${
+          response.status
+        }. Response body: ${stringify(response.data)}`
+      )
 
       throw e
     }