+++ homebridge/plugin-info.ts
@@ -1,4 +1,0 @@
-export const platformName = 'Ring'
-export const pluginName = 'homebridge-ring'
-export const oldPlatformName = 'RingAlarm'
-export const oldPluginName = 'homebridge-ring-alarm'